// src/models/auth/profile.interface.ts
export interface Profile {
    uid: string;
    displayName: string,
    email: string,
    emailVerified: boolean;
    phoneNumber: string;
    photoURL: string,
    companyname: string;
    address: string;
    street: string;
    mobile: string;
    city: string;
    country: string;
    workinghours: string;
    usertype: string;
    facebook: string;
    twitter: string;
    linkedin: string;
    googleplus: string;
    youtube: string;
    instagram: string;
}